test: separate mysql_repo, bolt_repo from repository.go
