+++ mysql_repository.go
@@ -1,0 +1,5 @@
+package main
+
+import (
+	_ "github.com/go-sql-driver/mysql"
+)
